+++ Covid19ModelVN/src/scripts/include/experiments.jl
@@ -58,7 +58,7 @@ function experiment_covid19_data(
     end
 
     df = get_prebuilt_covid_timeseries(loc)
-    df[!, cols] .= Float64.(df[!, cols])
+    df[!, cols] .= Float32.(df[!, cols])
     experiment_covid19_counts_reset!(df, loc)
     # choose the first date to be when the number of total confirmed cases passed 500
     first_date = first(subset(df, :confirmed_total => x -> x .>= 500, view = true).date)
@@ -88,7 +88,7 @@ function experiment_movement_range(
 )
     df = get_prebuilt_movement_range(loc)
     cols = ["all_day_bing_tiles_visited_relative_change", "all_day_ratio_single_tile_users"]
-    df[!, cols] .= Float64.(df[!, cols])
+    df[!, cols] .= Float32.(df[!, cols])
     # smooth out weekly seasonality
     if ma7
         moving_average!(df, cols, 7)
@@ -103,7 +103,7 @@ function experiment_social_proximity(
     ma7::Bool,
 )
     df, col = get_prebuilt_social_proximity(loc)
-    df[!, col] .= Float64.(df[!, col])
+    df[!, col] .= Float32.(df[!, col])
     # smooth out weekly seasonality
     if ma7
         moving_average!(df, col, 7)
@@ -158,7 +158,7 @@ function experiment_loss(
     weights = exp.(collect(tsteps) .* ζ)
     domain = max .- min
     lossfn = @inbounds function (ŷ, y)
-        s::Float64 = 0
+        s::Float32 = 0
         sz = size(y)
         for j ∈ 1:sz[2], i ∈ 1:sz[1]
             s += ((ŷ[i, j] - y[i, j]) / domain[i])^2 * weights[j]
@@ -169,14 +169,14 @@ function experiment_loss(
 end
 
 SEIRDBaselineHyperparams = @NamedTuple begin
-    L2_λ::Float64
-    ζ::Float64
-    γ0::Float64
-    λ0::Float64
-    α0::Float64
-    γ_bounds::Tuple{Float64,Float64}
-    λ_bounds::Tuple{Float64,Float64}
-    α_bounds::Tuple{Float64,Float64}
+    L2_λ::Float32
+    ζ::Float32
+    γ0::Float32
+    λ0::Float32
+    α0::Float32
+    γ_bounds::Tuple{Float32,Float32}
+    λ_bounds::Tuple{Float32,Float32}
+    α_bounds::Tuple{Float32,Float32}
     train_range::Day
     forecast_range::Day
     ma7::Bool
@@ -211,14 +211,14 @@ function setup_baseline(loc::AbstractString, hyperparams::SEIRDBaselineHyperpara
 end
 
 SEIRDFbMobility1Hyperparams = @NamedTuple begin
-    L2_λ::Float64
-    ζ::Float64
-    γ0::Float64
-    λ0::Float64
-    α0::Float64
-    γ_bounds::Tuple{Float64,Float64}
-    λ_bounds::Tuple{Float64,Float64}
-    α_bounds::Tuple{Float64,Float64}
+    L2_λ::Float32
+    ζ::Float32
+    γ0::Float32
+    λ0::Float32
+    α0::Float32
+    γ_bounds::Tuple{Float32,Float32}
+    λ_bounds::Tuple{Float32,Float32}
+    α_bounds::Tuple{Float32,Float32}
     train_range::Day
     forecast_range::Day
     ma7::Bool
@@ -263,14 +263,14 @@ function setup_fbmobility1(loc::AbstractString, hyperparams::SEIRDFbMobility1Hyp
 end
 
 SEIRDFbMobility2Hyperparams = @NamedTuple begin
-    L2_λ::Float64
-    ζ::Float64
-    γ0::Float64
-    λ0::Float64
-    α0::Float64
-    γ_bounds::Tuple{Float64,Float64}
-    λ_bounds::Tuple{Float64,Float64}
-    α_bounds::Tuple{Float64,Float64}
+    L2_λ::Float32
+    ζ::Float32
+    γ0::Float32
+    λ0::Float32
+    α0::Float32
+    γ_bounds::Tuple{Float32,Float32}
+    λ_bounds::Tuple{Float32,Float32}
+    α_bounds::Tuple{Float32,Float32}
     train_range::Day
     forecast_range::Day
     social_proximity_lag::Day
@@ -326,15 +326,15 @@ function setup_fbmobility2(loc::AbstractString, hyperparams::SEIRDFbMobility2Hyp
 end
 
 SEIRDFbMobility3Hyperparams = @NamedTuple begin
-    L2_λ::Float64
-    ζ::Float64
-    γ0::Float64
-    λ0::Float64
-    α0::Float64
-    β_bounds::Tuple{Float64,Float64}
-    γ_bounds::Tuple{Float64,Float64}
-    λ_bounds::Tuple{Float64,Float64}
-    α_bounds::Tuple{Float64,Float64}
+    L2_λ::Float32
+    ζ::Float32
+    γ0::Float32
+    λ0::Float32
+    α0::Float32
+    β_bounds::Tuple{Float32,Float32}
+    γ_bounds::Tuple{Float32,Float32}
+    λ_bounds::Tuple{Float32,Float32}
+    α_bounds::Tuple{Float32,Float32}
     train_range::Day
     forecast_range::Day
     social_proximity_lag::Day
@@ -391,14 +391,14 @@ function setup_fbmobility3(loc::AbstractString, hyperparams::SEIRDFbMobility3Hyp
 end
 
 SEIRDFbMobility4Hyperparams = @NamedTuple begin
-    L2_λ::Float64
-    ζ::Float64
-    γ0::Float64
-    λ0::Float64
-    β_bounds::Tuple{Float64,Float64}
-    γ_bounds::Tuple{Float64,Float64}
-    λ_bounds::Tuple{Float64,Float64}
-    α_bounds::Tuple{Float64,Float64}
+    L2_λ::Float32
+    ζ::Float32
+    γ0::Float32
+    λ0::Float32
+    β_bounds::Tuple{Float32,Float32}
+    γ_bounds::Tuple{Float32,Float32}
+    λ_bounds::Tuple{Float32,Float32}
+    α_bounds::Tuple{Float32,Float32}
     train_range::Day
     forecast_range::Day
     social_proximity_lag::Day
@@ -552,8 +552,8 @@ function experiment_run(
     savedir::AbstractString,
     kwargs...,
 )
-    minimizers = Vector{Float64}[]
-    final_losses = Float64[]
+    minimizers = Vector{Float32}[]
+    final_losses = Float32[]
     lk = ReentrantLock()
 
     run = function (loc)
